chore: 파일 구조 변경 (#58)
