+++ src/main/java/com/newfit/reservation/common/ScheduleService.java
@@ -1,4 +1,4 @@
-package com.newfit.reservation.service;
+package com.newfit.reservation.common;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -6,8 +6,10 @@ import java.util.List;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
-import com.newfit.reservation.common.fcm.dto.NotificationRequest;
-import com.newfit.reservation.common.fcm.service.FCMService;
+import com.newfit.reservation.domains.authority.domain.Authority;
+import com.newfit.reservation.domains.authority.repository.AuthorityRepository;
+import com.newfit.reservation.domains.fcm.dto.NotificationRequest;
+import com.newfit.reservation.domains.fcm.service.FCMService;
 import com.newfit.reservation.domains.reservation.domain.Reservation;
 import com.newfit.reservation.domains.reservation.repository.ReservationRepository;
 
@@ -16,8 +18,11 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class ScheduleService {
+
 	private final ReservationRepository reservationRepository;
 	private final FCMService fcmService;
+	private final AuthorityRepository authorityRepository;
+
 	private static final String NOTIFICATION_TITLE = "예약 5분 전 알림";
 
 	@Scheduled(cron = "0 0/1 * * * *")
@@ -30,5 +35,12 @@ public class ScheduleService {
 		for (NotificationRequest notificationRequest : notificationRequests) {
 			fcmService.sendMessageByToken(notificationRequest);
 		}
+	}
+
+	// 매일 자정에 모든 Authority의 크레딧 획득 횟수를 초기화
+	@Scheduled(cron = "0 0 0 * * ?")
+	private void resetCreditAcquisitionCount() {
+		List<Authority> authorities = authorityRepository.findAllByCreditAcquisitionCountNotZero();
+		authorities.forEach(Authority::resetAcquisitionCount);
 	}
 }
