--- src/main/java/com/newfit/reservation/common/fcm/dto/NotificationResponse.java
@@ -1,6 +1,0 @@
-package com.newfit.reservation.common.fcm.dto;
-
-public class NotificationResponse {
-    private int status;
-    private String message;
-}
