+++ src/main/java/com/newfit/reservation/domains/fcm/dto/NotificationResponse.java
@@ -1,0 +1,6 @@
+package com.newfit.reservation.domains.fcm.dto;
+
+public class NotificationResponse {
+	private int status;
+	private String message;
+}
