+++ src/main/java/com/newfit/reservation/domains/fcm/repository/FCMRepository.java
@@ -1,8 +1,8 @@
-package com.newfit.reservation.common.fcm.repository;
+package com.newfit.reservation.domains.fcm.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.newfit.reservation.common.fcm.domain.FCMToken;
+import com.newfit.reservation.domains.fcm.domain.FCMToken;
 
 public interface FCMRepository extends JpaRepository<FCMToken, Long> {
 }
