+++ src/main/resources/application.properties
@@ -3,19 +3,14 @@ spring.jpa.database=postgresql
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.use_sql_comments=true
 spring.jpa.defer-datasource-initialization=true
-
 #security
 spring.profiles.include=security
-
 #aws s3
 cloud.aws.region.static=ap-northeast-2
 cloud.aws.region.auto=false
 cloud.aws.stack.auto=false
 cloud.aws.s3.bucket=newfit-image
-
 #redis
 spring.data.redis.host=localhost
 spring.data.redis.port=6379
 
-#firebase fcm
-app.firebase-configuration-file=newfit-d92c2-firebase-adminsdk.json
